requirements.txt
# Dependencias principales
streamlit>=1.24.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.14.0

# Procesamiento de datos
scikit-learn>=1.2.0
nltk>=3.8.0
qdrant-client>=1.6.0
sentence-transformers>=2.2.2

# Integración con APIs
requests>=2.28.0
python-dotenv>=1.0.0
openai>=1.0.0
langchain>=0.0.267
langchain-openai>=0.0.5
huggingface-hub>=0.19.0
pygithub>=1.59.0

# Sincronización y programación
schedule>=1.2.0
gitpython>=3.1.30
apscheduler>=3.10.0

# Herramientas de desarrollo
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0

# Optimizaciones para Mac M2
# Estas dependencias están optimizadas para arquitecturas Apple Silicon
tensorflow-macos>=2.12.0; sys_platform == 'darwin' and platform_machine == 'arm64'
torch>=2.0.0; sys_platform == 'darwin'

# Documentación
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0